fix problem in QR-code data
If no extra_infos was given, the value used in the
actual QR-code was None. To work around this, use
or '' for all parameters which are not of type address

Somehow the problem does not show up in the tests, but
only in the QR-code picture itself, which can only be
seen when actually decoding the QR-code back to
textual form.