use python-stdnum instead of validators
In the validators module, only the iban part is relevant
to swiss-qr-bill. We need this functionality also for
the reference numbers (QRR / SCOR). python-stdnum has
all of it.

As some patches to python-stdnum which we need have not
yet been released, temporarily include the relevant files
from python-stdnum directly as module tmp_stdnum