Update bill.py
remove commented code

The intention was to keep the already working code there to
simply remove the comments signs again once this functionality
is changed from "for future use" to "can now be used" in the
specifications. But of course, it will be easy to get the lines back
from the git history and the code is cleaner without them.