err out when ultimate creditor is used
* err out when ultimate creditor is used

UltmtCdtr (ulimate creditor) is in the specs. However,
it is marked as 'for future use' and the fields flagged as
'do not fill in'. The online validator also rejects QR-codes
where these fields are not empty.
As it might be used in a future version of the standard,
do not remove the already working code, but rather return
an error if the structure is attempted to be used.